Move DataFile to own file, and rename to DatasetFile.
